// src/app/api/model/entityModelCVResponse.ts
/**
 * Personal Finance Management API
 *
 * Contact: dragisa@yahoo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExperienceResponse } from './experienceResponse';
import { SocialLinkResponse } from './socialLinkResponse';
import { EducationResponse } from './educationResponse';
import { Link } from './link';


export interface EntityModelCVResponse { 
    id?: number;
    username?: string;
    fullName?: string;
    email: string;
    phoneNumber?: string;
    summary?: string;
    address?: string;
    experiences?: Array<ExperienceResponse>;
    educations?: Array<EducationResponse>;
    socialLinks?: Array<SocialLinkResponse>;
    skills?: Array<string>;
    languages?: Array<string>;
    profileImage?: string;
    _links?: { [key: string]: Link; };
}